requirements.txt
greenlet==2.0.2
grpcio==1.56.0
grpcio-tools==1.56.0
protobuf==4.23.4
psycopg2==2.9.6
SQLAlchemy==2.0.19
typing_extensions==4.7.1
